fix(components): refactor search component
