+++ src/lib/layout/search/SearchComponent.tsx
@@ -34,55 +34,9 @@ import { useSearchHandler } from "lib/services/searchService";
 import type { Addr, Option } from "lib/types";
 import { splitModule } from "lib/utils";
 
-import { getRouteOptions, SearchResultRenderer } from "./SearchResultRenderer";
-
-const getZeroState = ({
-  isWasm,
-  isPool,
-  isMove,
-  isGov,
-  isFullTier,
-}: {
-  isWasm: boolean;
-  isPool: boolean;
-  isMove: boolean;
-  isGov: boolean;
-  isFullTier: boolean;
-}) => {
-  const starter = ["Account Address", "TX Hash", "Block Height"];
-  const govText = isGov ? ["Validator Address", "Proposal ID"] : [];
-  const wasmText = isWasm ? ["Code ID", "Contract Address"] : [];
-  const moveText = isMove ? ["Module Path"] : [];
-  const poolText = isPool && isFullTier ? ["Pool ID"] : [];
-
-  const supportedItemsType = starter.concat(
-    govText,
-    wasmText,
-    moveText,
-    poolText
-  );
-
-  return (
-    <Flex direction="column" gap={4} py={8}>
-      <Text color="text.dark"> Please enter keyword, You can search with:</Text>
-      <Flex
-        direction="column"
-        px={4}
-        py={2}
-        border="1px solid"
-        borderColor="gray.700"
-        borderRadius={8}
-      >
-        {supportedItemsType.map((item) => (
-          <Flex alignItems="center" gap={3}>
-            <Flex w={1} h={1} borderRadius="full" bgColor="primary.darker" />
-            <Text color="text.dark">{item}</Text>
-          </Flex>
-        ))}
-      </Flex>
-    </Flex>
-  );
-};
+import { SearchResultRenderer } from "./SearchResultRenderer";
+import { SearchZeroState } from "./SearchZeroState";
+import { getRouteOptions } from "./utils";
 
 const getNextCursor = (
   key: string,
@@ -377,7 +331,13 @@ export const SearchComponent = () => {
               </>
             ) : (
               <Flex justifyContent="center">
-                {getZeroState({ isWasm, isPool, isMove, isGov, isFullTier })}
+                <SearchZeroState
+                  isWasm={isWasm}
+                  isPool={isPool}
+                  isMove={isMove}
+                  isGov={isGov}
+                  isFullTier={isFullTier}
+                />
               </Flex>
             )}
           </ModalBody>
