+++ src/lib/layout/search/utils.ts
@@ -1,0 +1,41 @@
+import type { SearchResultType } from "lib/services/searchService";
+import type { Nullable, Option } from "lib/types";
+
+export const getRouteOptions = (
+  type: Option<SearchResultType>
+): Nullable<{ pathname: string; query: string[] }> => {
+  switch (type) {
+    case "Account Address":
+      return {
+        pathname: "/accounts/[accountAddress]",
+        query: ["accountAddress"],
+      };
+    case "Transaction Hash":
+      return { pathname: "/txs/[txHash]", query: ["txHash"] };
+    case "Code ID":
+      return { pathname: "/codes/[codeId]", query: ["codeId"] };
+    case "Contract Address":
+      return {
+        pathname: "/contracts/[contractAddress]",
+        query: ["contractAddress"],
+      };
+    case "Block":
+      return { pathname: "/blocks/[height]", query: ["height"] };
+    case "Proposal ID":
+      return { pathname: "/proposals/[proposalId]", query: ["proposalId"] };
+    case "Validator Address":
+      return {
+        pathname: "/validators/[validatorAddress]",
+        query: ["validatorAddress"],
+      };
+    case "Pool ID":
+      return { pathname: "/pools/[poolId]", query: ["poolId"] };
+    case "Module Path":
+      return {
+        pathname: "/modules/[address]/[moduleName]",
+        query: ["address", "moduleName"],
+      };
+    default:
+      return null;
+  }
+};
